Save firefly.log in kira.sh; don't create 'done'
Let the Makefile that calls kira.sh create it instead. This is
to make it similar to tempwrap.

+++ kira.sh
@@ -12,7 +12,7 @@
 # The Kira binary being executed is `$KIRA` if that environment
 # variable exists, otherwise just `kira`.
 
-if [ $# < 1 ]; then
+if [ $# -lt 1 ]; then
     echo "usage: $0 /path/to/jobs.yaml [kira-arguments] ..." 1>&2
     exit 1
 fi
@@ -32,17 +32,12 @@ tmpdir="$(mktemp -d)"
 cp -va "$yamldir/." "$tmpdir/" || exit 1
 (cd "$tmpdir" && ${KIRA:-kira} "$yamlbase" "$@")
 code="$?"
-if [ -e "$tmpdir/results" ]; then
-    rm -f "$tmpdir/results/kira.db"
-    cp -a "$tmpdir/results" "$yamldir/"
-fi
-if [ -e "$tmpdir/kira.log" ]; then
-    cp -a "$tmpdir/kira.log" "$yamldir/"
-fi
+[ -e "$tmpdir/results/kira.db" ] && rm -f "$tmpdir/results/kira.db"
+[ -e "$tmpdir/results" ] && cp -a "$tmpdir/results" "$yamldir/"
+[ -e "$tmpdir/kira.log" ] && cp -a "$tmpdir/kira.log" "$yamldir/"
+[ -e "$tmpdir/firefly.log" ] && cp -a "$tmpdir/firefly.log" "$yamldir/"
 rm -rf "$tmpdir"
-if [ "$code" -eq 0 ]; then
-    date > "$yamldir/done"
-else
+if [ "$code" -ne 0 ]; then
     echo "Error: ${KIRA:-kira} failed with exit code $code"
 fi
 exit "$code"
